@@ -1,4 +1,5 @@
 library cubos_widgets;
+
 import 'package:flutter/material.dart';
 
 /// Returns a SizedBox with the given [height]
@@ -6,8 +7,11 @@ import 'package:flutter/material.dart';
 /// It is very useful to set a margin between widgets in a Column.
 class VerticalSpacing extends StatelessWidget {
   final double height;
-
-  const VerticalSpacing(this.height): assert(height >= 0);
+  const VerticalSpacing({
+    Key? key,
+    required this.height,
+  })  : assert(height >= 0),
+        super(key: key);
 
   @override
   Widget build(BuildContext context) {
